src/test/java/com/kovalchuk/longest_palindromic_substring/SolutionTest.java
package com.kovalchuk.longest_palindromic_substring;

import org.junit.jupiter.api.BeforeAll;
import org.junit.jupiter.api.TestInstance;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

import static org.junit.jupiter.api.Assertions.assertEquals;

@TestInstance(TestInstance.Lifecycle.PER_CLASS)
class SolutionTest {
    private Solution solution;

    @BeforeAll
    public void init() {
        solution = new Solution();
    }

    static Stream<Arguments> testData() {
        return Stream.of(
                Arguments.of("babad", "bab"),
                Arguments.of("cbbd", "bb")
        );
    }

    @ParameterizedTest
    @MethodSource("testData")
    void longestPalindrome(String s, String expected) {
        String actual = solution.longestPalindrome(s);

        assertEquals(expected, actual);
    }
}